+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - added export setting "objectscript.export.addCategory" if enabled uses previous behavior, adds category folder to export folder, disabled by default
 - added Server actions menu, by clicking on server info from status bar. Open Management portal, Class Reference and toggle connection.
+- Class Suggestion in ##class, Extends, As, CompileAfter, DependsOn, PropertyClass
+- \$SYSTEM suggestion by Classes from %SYSTEM
 
 ## [0.7.10]